Move local components to independent files
